增加下单金额兑换成usdt总额

+++ src/main/java/com/money/game/robot/huobi/response/OrdersDetail.java
@@ -65,5 +65,5 @@ public class OrdersDetail implements Serializable{
     private String state;
 
     @com.google.gson.annotations.SerializedName("canceled_at")
-    private int canceledAt;
+    private long canceledAt;
 }
